fixed CQ for bitmessagekivy.mpybit module part8 along with platform separation for linux and android
